+++ tests/utils/test_linear_cross_entropy.py
@@ -111,7 +111,7 @@ class TestLinearCrossEntropy:
     def generate_hyper(self):
         global MAX_TEST_CASES
 
-        self.dtype = torch.bfloat16
+        self.dtype = torch.float16
         if self.test_case_idx == 0:
             self.batch_size = 1
             self.num_tokens = 1937
