+++ tests/workers/rollout/rollout_vllm/run_fsdp_vllm.py
@@ -42,7 +42,7 @@ def main():
     actor_model_config = AutoConfig.from_pretrained(local_model_path, trust_remote_code=True)
     with torch.device("cuda"):
         actor_model = AutoModelForCausalLM.from_pretrained(local_model_path, trust_remote_code=True)
-        actor_model.to(torch.bfloat16)
+        actor_model.to(torch.float16)
 
     max_prompt_length = 16
     response_length = 32
@@ -87,7 +87,7 @@ def main():
 
     device_mesh = init_device_mesh("cuda", mesh_shape=(world_size,), mesh_dim_names=["fsdp"])
 
-    mixed_precision = MixedPrecision(param_dtype=torch.bfloat16, reduce_dtype=torch.float32, buffer_dtype=torch.float32)
+    mixed_precision = MixedPrecision(param_dtype=torch.float16, reduce_dtype=torch.float32, buffer_dtype=torch.float32)
     fsdp_model = FSDP(
         actor_model,
         use_orig_params=True,
