+++ verl/model_merger/base_model_merger.py
@@ -218,7 +218,7 @@ class BaseModelMerger(ABC):
     def save_hf_model_and_tokenizer(self, state_dict: dict[str, torch.Tensor]):
         auto_model_class = self.get_transformers_auto_model_class()
         with init_empty_weights():
-            model = auto_model_class.from_config(self.model_config, torch_dtype=torch.bfloat16)
+            model = auto_model_class.from_config(self.model_config, torch_dtype=torch.float16)
         model.to_empty(device="cpu")
         model = self.patch_model_generation_config(model)
